add signup/login/logout
This adds some basic Svelte Kit hooks/endpoints to handle connecting
supabase's auth up to how Svelte Kit handles sessions. The purpose of
this is twofold:

  1. Read during SSR with the user's credentials
  2. Allow for "just a form" with progressive enhancement

The client picks up the access_token from document.cookie and passes it
to auth.setAuth(). This requires the cookies to omit HttpOnly, as that
will allow them to be readable in JS. I had to add a helper for
listening to changes on document.cookie, as window.cookieStore is brand
new and not available in all browsers.