performant rls solution for boards/completions
This is largely based on https://jeffchen.dev/posts/Postgres-13-Row-Level-Security-Performance/

The idea is to allow the query planner to optimize a bit further by not
including the value to check in the subquery as a clause. In other
words:

Don't:
	EXISTS(
		SELECT 1
		FROM games_users gu
		WHERE gu.game_id = _game_id AND gu.profile_id = profile_id()
	)

Because that needs a new query for every call. Instead,

Do:
SELECT _game_id = any(
	SELECT ARRAY(
		SELECT game_id FROM games_users WHERE profile_id = profile_id()
	)
)

This keeps the underlying set the same, and each call merely checks if
the game id exists within it. Note that this assumes a reasonable limit
for the number of games to which a user can belong.

For Boards and Completions, a function is used to return the
denormalized table I was discussing. Then each policy can use something
like this for read:

... USING (
	board_id IN (
		SELECT gcfu.board_id
		FROM get_completions_for_user() gcfu(board_id, ...)
	)
);

get_completions_for_user() is effectively every "tile" that the user can
see, so we might eventually look to pairing this down by game or simply
moving the denormalization from the function here to the actual schema
and just use it that way. However, for the time being, I wanted to
figure out how to make this work with the schema as is.